Extract the Oidc module (OIDC flow, discovery, token) into Api/Oidc
Tenth module of the #777 folder migration — the OIDC surface. Moves the 14 OIDC
types into Jellyfin.Plugin.SSO_Auth.Api.Oidc:

- OidcAuthorizeStateBuilder, OidcCallbackPath, OidcDiscovery{Options,Reader,
  Result}, OidcIdTokenValidator, OidcInsecureToggles, OidcResponseIssuer,
  OidcRoleExtractor, OidcStateStore, PkceDiscovery, DiscoveryFacts,
  AuthorizeSession, AuthorizeStateBinding -> Api/Oidc/ (namespace Api.Oidc).

Oidc orchestrates several lower modules: allowed deps are Authz, Avatar, Net,
Provider, RateLimit. Registers those edges.

The migration surfaced a real dependency CYCLE and broke it:
ProviderConnectionTester lived in the Provider module but imports Oidc (it
probes OIDC discovery) AND Saml (it probes SAML) — so Provider -> Oidc while
Oidc -> Provider. It is a cross-protocol orchestration type, not a Provider
primitive, so it moves back to the flat Api core to await the final Http/
orchestration module (which may depend on both protocol modules). Provider is
now clean of Oidc; no sibling module imports another.

Broken doc-cref references the move forced are fully-qualified rather than
resolved with a wrong-direction import: OidcCallbackPath/OidcStateStore/etc.
crefs in the Saml module and in flat-Api routing types (ChallengePath,
RouteSuffix), and Oidc's own cref to the now-flat ProviderConnectionTester.

Pure mechanical move, no behavior change. Verified by the full build
(--warnaserror, both TFMs), net9.0 tests (1580) + the module DAG conformance
rule, and the ABI-floor build. (net10.0 test execution was blocked locally by a
Windows application-control policy — a host issue, not a code one; CI runs the
net10.0 suite on Linux.)

Refs #777
