+++ SSO-Auth.Tests/AuthorizeStateBindingTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.Oidc;
 using Microsoft.AspNetCore.Http;
 using Xunit;
