+++ SSO-Auth.Tests/OidcResponseIssuerTests.cs
@@ -1,4 +1,5 @@
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.Oidc;
 using Microsoft.IdentityModel.Tokens;
 using Xunit;
