+++ SSO-Auth.Tests/PkceDiscoveryTests.cs
@@ -1,4 +1,5 @@
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.Oidc;
 using Xunit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Tests;
