+++ SSO-Auth/Api/Oidc/AuthorizeStateBinding.cs
@@ -2,7 +2,7 @@ using System;
 using System.Security.Cryptography;
 using Microsoft.AspNetCore.Http;
 
-namespace Jellyfin.Plugin.SSO_Auth.Api;
+namespace Jellyfin.Plugin.SSO_Auth.Api.Oidc;
 
 /// <summary>
 /// Binds an in-flight OpenID authorize state to the browser that started it (#326). The authorize
