+++ SSO-Auth/Api/Oidc/DiscoveryFacts.cs
@@ -1,4 +1,4 @@
-namespace Jellyfin.Plugin.SSO_Auth.Api;
+namespace Jellyfin.Plugin.SSO_Auth.Api.Oidc;
 
 /// <summary>
 /// The two discovery-document facts the OpenID challenge reads from a single discovery response: whether
