+++ SSO-Auth/Api/Oidc/OidcDiscoveryOptions.cs
@@ -3,14 +3,14 @@ using Duende.IdentityModel.OidcClient;
 using Jellyfin.Plugin.SSO_Auth.Api.Provider;
 using Jellyfin.Plugin.SSO_Auth.Config;
 
-namespace Jellyfin.Plugin.SSO_Auth.Api;
+namespace Jellyfin.Plugin.SSO_Auth.Api.Oidc;
 
 /// <summary>
 /// Builds the <see cref="OidcClientOptions"/> Authority and discovery policy — <c>RequireHttps</c>,
 /// <c>ValidateIssuerName</c>, <c>ValidateEndpoints</c>, and the additional base address for providers whose
 /// endpoints sit off the authority — from a provider config, in ONE place. Both the login path
 /// (<see cref="Flows.OidcLoginService"/>, which layers the client credentials, redirect URI, scope and the
-/// id_token validator on top) and the admin Test-connection probe (<see cref="ProviderConnectionTester"/>,
+/// id_token validator on top) and the admin Test-connection probe (<see cref="Jellyfin.Plugin.SSO_Auth.Api.ProviderConnectionTester"/>,
 /// which reads discovery only) build their options here, so the test fetch runs under the EXACT same
 /// SSRF/TLS posture as the real login discovery (#163) — a later change to the login's discovery policy
 /// cannot silently leave the probe on a weaker one. The client secret is deliberately NOT set here:
