+++ SSO-Auth/Api/Oidc/OidcRoleExtractor.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace Jellyfin.Plugin.SSO_Auth.Api;
+namespace Jellyfin.Plugin.SSO_Auth.Api.Oidc;
 
 /// <summary>
 /// Reads the role values out of an OpenID claim according to the configured role-claim path.
