+++ SSO-Auth/Api/Oidc/PkceDiscovery.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace Jellyfin.Plugin.SSO_Auth.Api;
+namespace Jellyfin.Plugin.SSO_Auth.Api.Oidc;
 
 /// <summary>
 /// Decides whether an OpenID provider's discovery document advertises PKCE with SHA-256 (<c>S256</c>).
