+++ SSO-Auth/Api/Saml/SamlAuthorizeStateBuilder.cs
@@ -12,7 +12,7 @@ namespace Jellyfin.Plugin.SSO_Auth.Api.Saml;
 /// only (roles, config) and returns the derived values, which the SAML callback applies to the
 /// session parameters. Login validity is decided separately by <see cref="SamlLoginPolicy"/> and the
 /// username is the assertion's NameID, so — unlike the OpenID builder
-/// (<see cref="OidcAuthorizeStateBuilder"/>) — neither is derived here. Mirrors, one-for-one, the
+/// (<see cref="Jellyfin.Plugin.SSO_Auth.Api.Oidc.OidcAuthorizeStateBuilder"/>) — neither is derived here. Mirrors, one-for-one, the
 /// privilege derivation that used to live inline in the callback.
 /// </summary>
 internal static class SamlAuthorizeStateBuilder
